+++ changelog.md
@@ -66,3 +66,6 @@
 - Added a character limit to the asset name and team name for the 29x90mm label
 - Added status to asset and kit
 - Updated logo and favicon
+
+## 0.3.13
+- Fixed QR Scanner to work with kits QR code